testing: Original conf file not recovered after tests
Original configuration file is not ripristinates anymore after the
end of testing. Currently the original configuration file is
stored in ram and it should be written again at the end of tests.

To avoid any loss, the original configuration file is now backed up
in a different location. If the issue happend again the original file
would be still available at /path/to/conf/file.bak