letsdo: Avoid time string in format like "N days"
Currently datetime.timedelta values can sum up to more than 24 hours.
When this happens, by default the string representation shows the
correspondent number of days, while it is better to keep the format H:M
even for a large number of hours.

To implement this feature I took a great contribute
from http://stackoverflow.com/a/42320260/1197008