report: Restore date value in report list
Currently the report shows only the amount of hours logged for each
task, while it is useful to show also the date of the last time the user
worked on each task.

This patch restore this information. In a next patch also the color
of the date shall be assigned