Alert when implicitly requesting report for --all
Executing letsdo without any flags defaults to using --all flag, that
might mean a long list of task done. Since this is a bit annoying, when
letsdo is run without any flag or search pattern, an alert is prompted
to inform the user about this default behaviour. The user can continue
pressing ENTER or Y if this is the right think or interrupt pressing N.