data report: review column orders
Review column order to make easier to check for a specific date.
Currently, the date of the task is after the description and since the
description has a variable length, the dates of a list of tasks are
almost never aligned, making it hard to search.

This change moves the date and work time after task ID, letting the
description - the only length variable part - at the end of the row